Adding: -Q02 Command Interpreter.cs -Q02 Problem Descroption.docx

+++ L11 Test/Test Preparation III/PT III/Q01 SoftUni Coffee/Program.cs
@@ -34,11 +34,11 @@ public class Program
         {
             var pricePerCapsule = decimal.Parse(Console.ReadLine());
             var date = DateTime.ParseExact(Console.ReadLine(), dateFormat, CultureInfo.InvariantCulture);
-            int numberOfCapsules = int.Parse(Console.ReadLine());
+            long numberOfCapsules = long.Parse(Console.ReadLine());
 
             int daysInMonth = DateTime.DaysInMonth(date.Year, date.Month);
 
-            decimal capsulesInMonth = daysInMonth * numberOfCapsules;
+            long capsulesInMonth = daysInMonth * numberOfCapsules;
 
             decimal totalForMonth = capsulesInMonth * pricePerCapsule;
